+++ components/nif/record.hpp
@@ -30,6 +30,7 @@ namespace Nif
 {
 
 class NIFFile;
+class NIFStream;
 
 enum RecordType
 {
@@ -97,7 +98,7 @@ struct Record
     Record() : recType(RC_MISSING), recIndex(~(size_t)0) {}
 
     /// Parses the record from file
-    virtual void read(NIFFile *nif) = 0;
+    virtual void read(NIFStream *nif) = 0;
 
     /// Does post-processing, after the entire tree is loaded
     virtual void post(NIFFile *nif) {}
