+++ components/nif/record_ptr.hpp
@@ -46,7 +46,7 @@ public:
     RecordPtrT() : index(-2) {}
 
     /// Read the index from the nif
-    void read(NIFFile *nif)
+    void read(NIFStream *nif)
     {
         // Can only read the index once
         assert(index == -2);
@@ -99,7 +99,7 @@ class RecordListT
     std::vector<Ptr> list;
 
 public:
-    void read(NIFFile *nif)
+    void read(NIFStream *nif)
     {
         int len = nif->getInt();
         list.resize(len);
